MAINT: Quick fix for the one pool world (#11)
This is just Wonchan's old PR with small comments added. I plan to open
an issue about actually adding at least some estimates about memory
usage. This may not matter for many simple pipelines, but also could
matter a lot for some.

(Memory use is not easy to bound, e.g. for joins or csv reading,
although even a very high overestimate may be useful if we read a small
csv for example.)

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Co-authored-by: Wonchan Lee <wonchanl@nvidia.com>