ENH: Support more return columns (#17)
Each column takes up some space in the return and this space is
currently limited but settable.

This bumps the limit for all tasks so that we should be able to use well
above 2500 columns everywhere.
However, using that many columns may require passing `--zcmem=100` or similar.

(Clearly, it is very plausible that we might bump this further if we run
into a use-case that needs that.)

---

Tested just with csv, since that is where it came up (but if used, it
tends to be the first call). Either way, it should affect everything, so
set it globally.

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>