ENH: Add ``apply_boolean_mask`` (#20)
Adding it with a similar file structure to libcudf which uses
``stream_compaction``.  (Not sure about the names, but also good to
keep aligned with libcudf.)

Straight forward exposure of the table-wise boolean masking.

---

Other fixup is due to sorting tests running into issues with the high
column things now, so lowered it.
(It must be unrelated, but the `test_task.py` seemed to hang when run as
an individual test.)

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Co-authored-by: Mads R. B. Kristensen <madsbk@gmail.com>