BUG: Sort code must deal with zero row local chunks (#21)
The local chunks may be empty and the sort code must make sure to deal
with it:
* An empty local table cannot share split points (the way it is done
here)
* As no split points are shared, the total number of split points may be
fewer than the number of ranks.

We need to deal with both of these. Now that there is boolean masking it
can be used to trigger these paths reliably (currently at least).

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>