BLD: Ensure we fetch the `_gpu` tag version (#25)
I am not sure if this should also be included in the dependencies.yaml,
but I suppose it shouldn't matter too much there for a release at least.
(conda didn't seem to like just adding it?)

This is identical to legate-boost (which needs to pick cpu or gpu
depending on the build).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>