ENH: Support `names=` in csv reading to not provide a header (#30)
We still don't allow usecols just being indices (we could).

In general, I would like to move to use csv_options at some point (even
if we have to copy out the ones we can deal with), but that might take a
bit and it seemed that this is a bit of a pain point.

TBH, I am a bit confused about the need for sorting, but it seems I also
was confused about it previously (i.e. further down there is another
sort).

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>