API: Introduce scalar columns (#36)
We don't want to deal with specialized scalar objects, but scalars are
special at least in that they are always length 1.

We could always allow broadcasting length 1 columns (allow them as
scalar). However, checking the number of rows is blocking if we use
unbound stores (which we normally do).
This is undesireable as task launch (e.g. for binary ops) may need to be
different (broadcast scalars).

General notes:
* `replace_nulls` usually doesn't need to return a nullable result but
without blocking on the scalar we don't know that for sure. (Not a
regression, but it may be nice to make an exception here.)
* I decided not to propagate "scalar" in e.g. `empty_like` by default.
It seemed simpler to err on the "column result" side and lose the scalar
information unless explicitly preserved.
* This does enable "scalars" in unary/binary ops.
* It means we implicitly allow "scalars" instead of columns by default.
That seems mostly fine to me (one could reject explicitly if needed).

Note that this PR does not optimize scalar handling in any way. It may
make sense to specialize unpacking/packing the scalar in the
PhysicalColumn.

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>