ENH: Implement basic reductions (#32)
This implements basic aggregations using a similar pattern to the simple
aggs/finalizer used in parts of cudf.

Since that pattern doesn't take into account the initial value and is in
the `detail` API, opted instead to create a custom helper.

That is, we use `(dtype, agg, initial)` to identify each reduce and then
split the reduce into two steps:
1. A first-pass that reduces only locally.
2. A finalize pass that works with local results or other final
   results.

Supporting further aggregations should be straight forward additions in
the future.

The design should generalize relatively well to re-use for other
aggregations (e.g. maybe to improve groupby-aggs) or to calculate
multipe aggregations for the same column (since e.g. many need the sum
or count temporary result).

The one thing that should be improved is that "count valid" is
relatively special and it should possibly become a specialized task and
be exposed as explicit API (with `count_nulls()`, we actually should
have both probably because `num_rows()` is blocking).

---

I'll note that I did see occasional hangs depending on how tests are run
(multi-gpu). I suspect they are identical to earlier hangs I have seen
that seem to originate in legate/legion.

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Co-authored-by: Mads R. B. Kristensen <madsbk@gmail.com>