API: Add `to_array()` methods to convert column or tables to arrays (#37)
I opted for `to_array()` because `to_cupynumeric()` just seems very
long, but happy to change!
(But happy to change that, I am not sure what I think is better.)

It may be that, especially for columns, `cupynumeric.asarray(col)`
should just be the long-term solution.

For tables, I am personally not sure: `pandas` does it, but in practice
implicitly converting columns into a new dimension isn't that obvious to
me. Tables and arrays are very different.

(Technically breaking, because previously converting times was somewhat
sensible, and now it is reject unless you use `get_logical_array()`
explicitly.)

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Co-authored-by: Mads R. B. Kristensen <madsbk@gmail.com>