MAINT: Reduce the number of scalars passed to tasks
This packs scalars into vectors.  Legate (now?) supports
`std::vectors<type>` just fine for the most part, but since it
unpacks them as a `span` (which is probably good), opted to keep
the old pattern for now.

Unfortunatley, `bool` and `string` vectors are not supported directly
so this uses overloads.

I did this hoping to resolve some issues that, in the end, didn't
resolve.  But while it is a bit tedious with the table, it seems
like a step in the right direction anyway.
(We should probably use the `span`, vector as scalar directly in
some places, and teach `legate` to to deal with bool and string
vectors.)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>