Update to legate 25.05 and align task config (#46)
This lists all the tasks explicitly now, the registrar pattern is
probably salvangable but I am not sure that legate wants to. (At which
point we could vendor it, but maybe this isn't so bad.)

---

We probably should just put this in soon... Although, there is a very
good chance that we either can revert most of the changes again or need
to restructure them differently due to deprecations.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>