API: Allow parquet reading into (transposed) array (#48)
This adds a call to do a chunked copy directly into a transposed array.
To be honest, the only reason to do this inside legate-df, is that we
have the cudf dependency here anyway.
To some degree, the reason is just to work-around current legate issues
with too many columns.

The transposed copy is modeled after cudf interleaved copying code.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>