BUG: Fix offset for and add sanity check to maybe_bind_buffer (#60)
Since it is currently called from outside tasks, this lingering bug
probably has no effect (the offset is always 0).

But this was a useful pattern when allowing eager allocations more
generally and writing output in gh-57 where this then triggers.

So let's just fix it, even if it may not ever be hit without gh-57.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>