ENH: Read parquet files by row group (#59)
Reorganize parquet reading into reading by row-group, which seems the
right way to chunk parquet file reading.  Hopefully(!) fine enough to read
quickly, but more fine grained is just slow and painful.

---

This is a bit of a story to get there with two steps forward and one
backwards.

I would like to use image constraint, but legate does not seem to
properly support them here (I have no idea why).
Thus, this goes "back" to using a late-bound store (half the idea of a
"read array" was to NOT do this but there is no helping it).

I did keep a decent chunk of the image constraint approach around even
if unused.

Which also means having to refactor/duplicate some of the copying code
because we have to work with buffers and can't use a `write_accessor` on
late bound stores.

This should make reading parquet files faster and more robust but of
course limits the partitioning to row-group multiples, which is probably
good for temproraries, but if row groups are huge and don't fit into
GPUs evently, it isn't necessarily ideal.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>