ENH: Work towards polars support (#65)
Note that this is an experimental try, and the may change the design significantly to e.g. realign again with cudf-polars. This would also move towards a `collect(engine=)` path (from a user perspective), which should be the long term goal either way.

This is based on `cudf-polars` mostly copied verbatim with minor
adjustments, changes are mostly e.g. confined to `do_evaluate`.

Should actually support much of what we need:
* polars reading
* basic CSV reading
* Binary, unary ops, and casts
* joins
* boolean filters.

Not yet, reductions or groupby-aggs and some other things. And there are
most likely many holes.

---

Right now (I need to see if there is a nicer way).  The main API is:
```
import legate_dataframe.ldf_polars  # needed to monkey-patch for now

query = # polars lazy frame query
query.legate.collect()  # query.collect(engine=...) can probably be done
```
I also added:
```
LogicalTable.lazy()
```
to go from legate-df to a polars `LazyFrame`.
(That part may need a lot of polishing, also since `engine=` may be an
alternative path, but if we want to return legate-df tables, this may
well be the only good way to do it.)

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>