Add (very) basic polars reductions
This is more of a start to have some very basic reductio.  Right now
many important ones are missing for various reasons.
min/max should almost work, but should deal with NaNs and any/all similarly
need to deal with nulls (default logic is different), and that is the same
in normal reductions, though.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>