Support logical operators (again) and in_between for polars (#81)
This re-adds supports for logical operators, which were missing because
arrow seems to require boolean inputs (i.e. an explicit cast).

Additionally, since `in_between` seems commonly used in polars ad
support for that again.
(Polars is a bit fun, in that it also only defines bitwise and/or it
seems.)

With this, I can run pdsh query 6 (with file limitations).

N.B.: Next steps for pdsh is:
1. Support of multiple files rather than just glob (I am tempted to do
the `glob.glob()` part only in Python and change the C++ API to require
a vector of files (or single file maybe, although `{filename}` is not
that bad).
2. Most queries require groups-by aggs.
3. (depending on how data is generated, data often contains decimals,
support for which needs to be added in some paths, but testing with
non-decimal files is fine)

## Checklist
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.
- [x] Run `./build.sh test` for local testing, see
[CONTRIBUTING.md](https://github.com/rapidsai/legate-dataframe/blob/main/CONTRIBUTING.md#testing).
- [x] Ensure `pre-commit` checks are passing see
[CONTRIBUTING.md](https://github.com/rapidsai/legate-dataframe/blob/main/CONTRIBUTING.md#Pre-commit-hooks).

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>