Also add support for parquet and eager alloc flag (and fix CPU paths) (#84)
Legate has some issues around image constraints, which means we have to
very explicitly set them for all involved variables (most annoyingly the
mask, I didn't check if that may be enough).

With that, the constraints work though.  This also fixes some CPU paths.

This is a follow-up to gh-57, because I didn't want that to bloat even
more with such hacks (in the end, the hack isn't too bad, though).

---

The rects were set up for the array case, but not actually used (because
2-D is still broken anyway). But really, legate should probably support
the 1-D setup also for the 2-D case.

Test locally with multiple GPUs with a number where the image constraint
matters.

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>