BUG: `LogicalTable.to_array()` should use the column `to_array()` (#86)
We need to use the `to_array()` function to work around the nullability
issue and add the convenience of allowing epxort when there are no nulls
(by doing an actual count if needed).

The initial/old version was missing this.

---

Admittedly, the test isn't _quite_ spot-on, because I think the
conversion from arrow will make the column non-nullable (seeing there
are no nulls). So only the second part really tests things (by checking
for the custom error that comes from counting).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>